chore: reorganize unit test contracts
