+++ tests/unit_test/pipeline/test_scheduler.py
@@ -1,0 +1,45 @@
+# SPDX-License-Identifier: Apache-2.0
+
+from __future__ import annotations
+
+from sglang_omni_v1.scheduling.messages import IncomingMessage
+from sglang_omni_v1.scheduling.simple_scheduler import SimpleScheduler
+from tests.unit_test.pipeline.helpers import run_scheduler
+
+
+def test_simple_scheduler_batch_and_error_contracts() -> None:
+    """Preserves batched success output and per-request batch failure emission."""
+    good = SimpleScheduler(
+        lambda payload: payload,
+        batch_compute_fn=lambda payloads: [payload.upper() for payload in payloads],
+        max_batch_size=2,
+        max_batch_wait_ms=10,
+    )
+    outputs = run_scheduler(
+        good,
+        [
+            IncomingMessage("req-1", "new_request", "a"),
+            IncomingMessage("req-2", "new_request", "b"),
+        ],
+        output_count=2,
+    )
+    assert {out.data for out in outputs} == {"A", "B"}
+
+    bad = SimpleScheduler(
+        lambda payload: payload,
+        batch_compute_fn=lambda payloads: ["only-one"],
+        max_batch_size=2,
+        max_batch_wait_ms=10,
+    )
+    outputs = run_scheduler(
+        bad,
+        [
+            IncomingMessage("req-1", "new_request", "a"),
+            IncomingMessage("req-2", "new_request", "b"),
+        ],
+        output_count=2,
+    )
+    assert {out.request_id for out in outputs} == {"req-1", "req-2"}
+    assert all(
+        out.type == "error" and isinstance(out.data, ValueError) for out in outputs
+    )
