+++ tests/unit_test/qwen3_omni/test_pipeline.py
@@ -2,18 +2,11 @@
 
 from __future__ import annotations
 
-from collections import deque
 from types import SimpleNamespace
 
-import numpy as np
 import pytest
 import torch
 
-from sglang_omni_v1.model_runner.thinker_model_runner import ThinkerModelRunner
-from sglang_omni_v1.models.qwen3_omni.components.code2wav_scheduler import (
-    Code2WavScheduler,
-)
-from sglang_omni_v1.models.qwen3_omni.components.talker import Qwen3OmniTalker
 from sglang_omni_v1.models.qwen3_omni.config import (
     Qwen3OmniPipelineConfig,
     Qwen3OmniSpeechPipelineConfig,
@@ -24,10 +17,7 @@ from sglang_omni_v1.models.qwen3_omni.request_builders import (
     build_sglang_thinker_request,
     project_preprocessing_to_mm_aggregate,
 )
-from sglang_omni_v1.models.qwen3_omni.talker_model_runner import QwenTalkerModelRunner
-from sglang_omni_v1.models.qwen3_omni.talker_scheduler import QwenTalkerScheduler
 from tests.unit_test.fixtures.qwen_fakes import (
-    FakeCode2WavModel,
     FakeQwenTokenizer,
     make_qwen_payload,
     make_qwen_state,
@@ -35,6 +25,7 @@ from tests.unit_test.fixtures.qwen_fakes import (
 
 
 def test_qwen_pipeline_config_and_state_contracts() -> None:
+    """Preserves Qwen text/speech topology and PipelineState coercion behavior."""
     text_config = Qwen3OmniPipelineConfig(model_path="model")
     speech_config = Qwen3OmniSpeechPipelineConfig(model_path="model")
 
@@ -66,6 +57,7 @@ def test_qwen_pipeline_config_and_state_contracts() -> None:
 
 
 def test_qwen_mm_aggregate_keeps_lightweight_inputs_and_prunes_after_merge() -> None:
+    """Preserves lightweight fan-in payloads and prunes consumed encoder tensors."""
     state = make_qwen_state(
         mm_inputs={
             "image": {
@@ -104,6 +96,7 @@ def test_qwen_mm_aggregate_keeps_lightweight_inputs_and_prunes_after_merge() -> 
 
 
 def test_qwen_thinker_request_and_decode_contracts() -> None:
+    """Preserves incremental text deltas, replacement-char suppression, and final text."""
     stream_state = PipelineState()
     tokenizer = FakeQwenTokenizer(pieces={1: "A", 2: "\ufffd", 3: "B"})
     first = list(
@@ -142,6 +135,7 @@ def test_qwen_thinker_request_and_decode_contracts() -> None:
 def test_qwen_sglang_request_hashes_media_tokens_without_changing_mrope_ids(
     monkeypatch: pytest.MonkeyPatch,
 ) -> None:
+    """Preserves hashed media pad tokens while M-RoPE still sees original ids."""
     captured: dict[str, torch.Tensor] = {}
 
     def fake_mrope(input_ids, model_inputs, thinker_config):
@@ -188,118 +182,3 @@ def test_qwen_sglang_request_hashes_media_tokens_without_changing_mrope_ids(
     assert pad_values["audio"] >= 256
     assert int(req_data.input_ids[1]) == pad_values["audio"]
     assert captured["input_ids"].tolist() == input_ids.tolist()
-
-
-def test_qwen_talker_feedback_fifo_and_stream_done_contract() -> None:
-    sched_req = SimpleNamespace(
-        data=SimpleNamespace(
-            pending_feedback_queue=deque([torch.tensor([1.0, 2.0])]),
-            pending_text_queue=deque(),
-            tts_pad_embed=torch.tensor([7.0, 8.0]),
-            thinker_chunks_done=False,
-        )
-    )
-
-    assert (
-        QwenTalkerModelRunner._take_next_decode_input_embed(
-            sched_req=sched_req,
-            device=torch.device("cpu"),
-            dtype=torch.float32,
-        )
-        is None
-    )
-    sched_req.data.pending_text_queue.append(torch.tensor([20.0, 20.0]))
-    assert torch.equal(
-        QwenTalkerModelRunner._take_next_decode_input_embed(
-            sched_req=sched_req,
-            device=torch.device("cpu"),
-            dtype=torch.float32,
-        ),
-        torch.tensor([21.0, 22.0]),
-    )
-
-    scheduler = object.__new__(QwenTalkerScheduler)
-    req_data = SimpleNamespace(
-        pending_text_queue=deque([torch.tensor([11.0, 12.0])]),
-        thinker_chunks_done=True,
-    )
-    payload = SimpleNamespace(
-        prefetched_chunks=[SimpleNamespace(data=torch.tensor([20.0, 20.0]))],
-        prefetched_stream_done=True,
-    )
-    assert scheduler._is_request_build_ready(payload, pending_stream_done=True)
-    scheduler._initialize_request_stream_state(req_data, payload)
-    assert len(req_data.pending_text_queue) == 1
-    assert torch.equal(req_data.pending_text_queue[0], torch.tensor([11.0, 12.0]))
-
-
-def test_qwen_code2wav_streams_incrementally_and_abort_clears_state() -> None:
-    model = FakeCode2WavModel(total_upsample=2)
-    scheduler = Code2WavScheduler(
-        model,
-        device="cpu",
-        stream_chunk_size=2,
-        left_context_size=1,
-        sample_rate=24000,
-    )
-    scheduler._payloads["req-1"] = make_qwen_payload(request_id="req-1")
-    scheduler._ensure_request_state("req-1")
-
-    scheduler._on_chunk("req-1", SimpleNamespace(data=torch.tensor([1, 10])))
-    scheduler._on_chunk("req-1", SimpleNamespace(data=torch.tensor([2, 20])))
-    scheduler._on_chunk("req-1", SimpleNamespace(data=torch.tensor([3, 30])))
-    scheduler._on_done("req-1")
-
-    message = scheduler.outbox.get_nowait()
-    audio = np.frombuffer(message.data.data["audio_waveform"], dtype=np.float32)
-    assert model.calls == [(1, 2, 2), (1, 2, 2)]
-    assert audio.shape == (6,)
-
-    scheduler._payloads["req-2"] = make_qwen_payload(request_id="req-2")
-    scheduler._ensure_request_state("req-2")
-    scheduler._pending_done.add("req-2")
-    scheduler.abort("req-2")
-    assert "req-2" not in scheduler._code_chunks
-    assert "req-2" not in scheduler._payloads
-    assert "req-2" not in scheduler._pending_done
-
-
-def test_qwen_model_runner_and_code_predictor_tensor_contracts() -> None:
-    class RecordingEmbed:
-        num_embeddings = 10
-
-        def __init__(self) -> None:
-            self.seen: torch.Tensor | None = None
-
-        def __call__(self, input_ids: torch.Tensor) -> torch.Tensor:
-            self.seen = input_ids.clone()
-            return torch.zeros((input_ids.shape[0], 4), dtype=torch.float32)
-
-    runner = ThinkerModelRunner.__new__(ThinkerModelRunner)
-    runner._embed_tokens = RecordingEmbed()
-    runner._image_token_id = 5
-    runner._video_token_id = 6
-    runner._audio_token_id = 7
-    req = SimpleNamespace(
-        omni_model_inputs={
-            "audio_embeds": torch.tensor([[1.0, 2.0, 3.0, 4.0]]),
-            "pad_values": {"audio": 999},
-        },
-        _omni_consumed=None,
-        is_chunked=0,
-    )
-    input_embeds, _, _ = runner._inject_multimodal_embeds(
-        SimpleNamespace(input_ids=torch.tensor([1, 999, 2]), extend_seq_lens_cpu=[3]),
-        SimpleNamespace(reqs=[req]),
-    )
-
-    assert (
-        int(runner._embed_tokens.seen.max().item())
-        < runner._embed_tokens.num_embeddings
-    )
-    assert torch.equal(input_embeds[1], torch.tensor([1.0, 2.0, 3.0, 4.0]))
-
-    logits = torch.tensor([[[0.0, 1.0, 2.0]], [[2.0, 1.0, 0.0]]])
-    sampled = Qwen3OmniTalker._sample_code_predictor_token(logits)
-    assert sampled.shape == (2, 1)
-    assert sampled[:, 0].tolist() == [2, 0]
